Don't filter example tests based on environment.
Travis, AppVeyor, and CIBot all fail when there  are breaking
changes between LineSearches and downstream packages, such as Optim,
that are also used in the examples.
Instead of filtering tests based on CI environment, we'll just have to
remember to add the tests back when the most recent downstream
application works.